добавить авторизацию и регистрацию

+++ controllers/cards.js
@@ -2,6 +2,7 @@ const mongoose = require('mongoose');
 const Card = require('../models/card');
 const NotFoundError = require('../errors/NotFoundError');
 const BadRequestError = require('../errors/BadRequestError');
+const ForbiddenError = require('../errors/ForbiddenError');
 
 module.exports.getAllCards = (req, res, next) => {
   Card.find({})
@@ -14,25 +15,36 @@ module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
-    .then((card) => res.send(card))
     .catch((err) => {
       if (err instanceof mongoose.Error.ValidationError) {
         throw new BadRequestError('Переданы некорректные данные');
       }
     })
+    .then((card) => res.send(card))
     .catch(next);
 };
 
 module.exports.deleteCard = (req, res, next) => {
-  Card.findOneAndDelete(req.params.cardId)
-    .then((card) => res.send({ data: card }))
+  Card.findById(req.params._id)
+    .orFail()
+    .catch(() => {
+      throw new NotFoundError({ message: 'Карточка не найдена' });
+    })
+    .then((card) => {
+      if (card.owner.toString() !== req.user._id) {
+        throw new ForbiddenError({ message: 'Доступ запрещен' });
+      }
+      Card.findOneAndDelete(req.params.cardId)
+        .then((item) => res.send({ data: item }));
+    })
     .catch(next);
 };
 
 module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
-    .orFail(() => {
-      throw new NotFoundError('Карточка не найдена');
+    .orFail()
+    .catch(() => {
+      throw new NotFoundError({ message: 'Карточка не найдена' });
     })
     .then((likes) => res.send({ data: likes }))
     .catch(next);
@@ -40,8 +52,9 @@ module.exports.likeCard = (req, res, next) => {
 
 module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
-    .orFail(() => {
-      throw new NotFoundError('Карточка не найдена');
+    .orFail()
+    .catch(() => {
+      throw new NotFoundError({ message: 'Карточка не найдена' });
     })
     .then((likes) => res.send({ data: likes }))
     .catch(next);
